Refactor total_items parameter to numeric type and add support for most expensive target selection in optimisation functions

+++ R/dmd_dose_optimise.R
@@ -534,7 +534,7 @@ dmd_dose_cost <- function(
     dose_delivered = numeric(),
     dose_delivered_unit = character(),
     over_delivery = numeric(),
-    total_items = integer(),
+    total_items = numeric(),
     cost_prorata_pence = numeric(),
     cost_whole_pack_pence = numeric(),
     dose_cost_pence = numeric(),
